Moves some framework functionality out of bootstrap.php
The idea behind this is that /config/bootstrap.php should hold a
minimum of things to get the application and environment in a state
that it can handle a request. Features provided by Labrador, like the
guide and development tools, should be handled as middleware to allow
users to easily remove those features as needed.

tl;dr

/config/bootstrap.php is for critical steps in getting the app setup.
Features belong in middleware.