Fix schemas symlink (#80)
Jekyll has issues with symlinks that exist outside of the root directory
for your site. This fix simply copies over the schemas from the
resources/schemas directory into docs/schemas. A better solution should
be developed to not have duplicate files but it is desired to have these
schemas exist at the URL we use to define them.