Add application start, stop, and state (#90)
In practice it may be necessary or desirable to explicitly stop the
Application from running; e.g. when your app is a long-running server
and it needs to be bounced or shut down. This PR refactors the
Application interface to have explicit start() and stop() methods
designed to make it more semantic for starting your app and explicitly
stopping it.