Ensure Plugin loading always occurs
Previously we were skipping over the PluginManager::loadPlugins process
if there were no registered Plugins. However, this could cause an
erroneous error if after the application has started the
PluginManager::getLoadedPlugins method is called. Intuitively this
should work and simply return an empty array since nothing was loaded.
However, since the loading process was never marked as having been ran
this method was throwing an error even though in the lifecycle of the
application it should not have been. This patch updates the plugin
loading process to ensure it is always marked as happened even if we
skip any actual Plugin loading.

Fixes #109